+++ frontend/fingertips-frontend/components/views/OneIndicatorOneAreaView/index.tsx
