+++ frontend/fingertips-frontend/components/views/OneIndicatorTwoOrMoreAreasView/index.tsx
