+++ frontend/fingertips-frontend/components/views/TwoOrMoreIndicatorsAreasView/index.tsx
