+++ frontend/fingertips-frontend/components/views/TwoOrMoreIndicatorsEnglandView/index.tsx
